Dry-Up some constants and re-organize config and config test code
